EditTripForm update trip onBlur
